# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vensle API ",
        "description": "Description of your API",
        "contact": {
            "email": "hybiekay2@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/auth/google/callback": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Handle Google authentication callback",
                "operationId": "46f883c727b554032ac992a94db80b84",
                "responses": {
                    "302": {
                        "description": "Redirect to the frontend with authentication token",
                        "content": {
                            "text/html": {
                                "example": "Redirect to frontend URL with authentication token"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}